requirements/common.txt
# Common dependencies
sanic==21.3.*
aiohttp==3.7.*
kafka-python==2.0.*